+++ packages/playground/src/App.tsx
@@ -1,13 +1,10 @@
-/* eslint-disable @typescript-eslint/dot-notation */
 import Editor from "@monaco-editor/react";
 import endent from "endent";
-// eslint-disable-next-line import/no-extraneous-dependencies -- we will use shared Prettier install from the root
 import * as prettier from "prettier";
 import * as prettierApex from "prettier-plugin-apex";
 import { useEffect, useState } from "react";
 import { useDebounce } from "use-debounce";
 
-// eslint-disable-next-line import/no-relative-packages -- this is easiest way to get Prettier Apex version
 import { version } from "../../prettier-plugin-apex/package.json";
 import icon from "../static/icon.png";
 import { ClipboardButton } from "./Buttons.js";