+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Fix trailing empty line not printed after ignored nodes in class declaration blocks ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/1892)).
 
+## Internal Changes
+
+- Replace `eslint` with `biome` for linting purpose.
+
 # 2.2.6
 
 - Fix dangling comments in `continue`/`break`/`return` statements not getting printed ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/1946)).